backend/requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.7

# Web Server (Webhook)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.10

# Database
aiosqlite==0.19.0

# Configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Logging
loguru==0.7.2
structlog==24.4.0

# Caching
cachetools==5.3.2
aiocache==0.12.3

# Address Validation
base58==2.1.1
eth-hash[pycryptodome]==0.5.2

# Blockchain (Contract Analysis)
web3==7.6.0
solana==0.35.0
solders==0.21.0

# Utilities
tenacity==9.0.0

# Security
slowapi==0.1.9
